moved navigation from App.jsx to Header component
